project/storage/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
pydantic
sqlalchemy
databases[postgresql]
python-multipart
psycopg2-binary  # ถ้าใช้ databases ไม่จำเป็นต้องใช้ psycopg2-binary
cors